internal/helpers/kube_helper.go
package helpers

import (
	"fmt"

	"github.com/compose-spec/compose-go/types"
	"github.com/windsor-hotel/cli/internal/context"
	"github.com/windsor-hotel/cli/internal/di"
)

// KubeHelper is a helper struct that provides Kubernetes-specific utility functions
type KubeHelper struct {
	Context context.ContextInterface
}

// NewKubeHelper is a constructor for KubeHelper
func NewKubeHelper(di *di.DIContainer) (*KubeHelper, error) {
	resolvedContext, err := di.Resolve("contextHandler")
	if err != nil {
		return nil, fmt.Errorf("error resolving context: %w", err)
	}

	return &KubeHelper{
		Context: resolvedContext.(context.ContextInterface),
	}, nil
}

// Initialize performs any necessary initialization for the helper.
func (h *KubeHelper) Initialize() error {
	// Perform any necessary initialization here
	return nil
}

// GetComposeConfig returns the top-level compose configuration including a list of container data for docker-compose.
func (h *KubeHelper) GetComposeConfig() (*types.Config, error) {
	// Stub implementation
	return nil, nil
}

// WriteConfig writes any vendor specific configuration files that are needed for the helper.
func (h *KubeHelper) WriteConfig() error {
	return nil
}

// Up executes necessary commands to instantiate the tool or environment.
func (h *KubeHelper) Up(verbose ...bool) error {
	return nil
}

// Info returns information about the helper.
func (h *KubeHelper) Info() (interface{}, error) {
	return nil, nil
}

// Ensure KubeHelper implements Helper interface
var _ Helper = (*KubeHelper)(nil)
